[UI] Move bug summary to generic helpers
